api final tcc

+++ src/main/java/tasknavigation/demo/repository/TarefaRepository.java
@@ -3,5 +3,5 @@ package tasknavigation.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tasknavigation.demo.domain.Tarefa;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Integer> {
+public interface TarefaRepository extends JpaRepository<Tarefa, Long> {
 }
